Move categorize email to its own file
